update hma to have static and class method
