+++ sdv/multi_table/hma.py
@@ -33,6 +33,126 @@ class HMASynthesizer(BaseHierarchicalSampler, BaseMultiTableSynthesizer):
         'default_distribution': 'beta'
     }
 
+    @staticmethod
+    def _get_root_tables(relationships):
+        parent_tables = {rel['parent_table_name'] for rel in relationships}
+        child_tables = {rel['child_table_name'] for rel in relationships}
+        return parent_tables - child_tables
+
+    @staticmethod
+    def _get_num_data_columns(metadata):
+        """Get the number of data columns, ie colums that are not id, for each table.
+
+        Args:
+            metadata (MultiTableMetadata):
+                Metadata of the datasets.
+        """
+        columns_per_table = {}
+        for table_name, table in metadata.tables.items():
+            columns_per_table[table_name] = \
+                sum([1 for col in table.columns.values() if col['sdtype'] != 'id'])
+
+        return columns_per_table
+
+    @classmethod
+    def _get_num_extended_columns(cls, metadata, table_name,
+                                  parent_table, columns_per_table, distributions=None):
+        """Get the number of columns that will be generated for table_name.
+
+        A table generates, for each foreign key:
+            - 1 num_rows column
+            - n*(n-1)/2 correlation columns for each data column
+            - k parameter columns for each data column, where:
+                - k = 4 if the distribution is beta or truncnorm (params are a, b, loc, scale)
+                - k = 3 if the distribution is gamma (params are a, loc, scale)
+                - k = 2 if the distribution is norm or uniform (params are loc, scale)
+        """
+        if distributions is None:
+            distribution = cls.DEFAULT_SYNTHESIZER_KWARGS['default_distribution']
+        else:
+            distribution = distributions.get(
+                table_name, cls.DEFAULT_SYNTHESIZER_KWARGS['default_distribution']
+            )
+
+        if distribution in {'beta', 'truncnorm'}:
+            num_parameters = 4
+        elif distribution == 'gamma':
+            num_parameters = 3
+        elif distribution in {'norm', 'uniform'}:
+            num_parameters = 2
+
+        num_rows_columns = len(metadata._get_foreign_keys(parent_table, table_name))
+
+        # no parameter columns are generated if there are no data columns
+        num_data_columns = columns_per_table[table_name]
+        if num_data_columns == 0:
+            return num_rows_columns
+
+        num_parameters_columns = num_rows_columns * num_data_columns * num_parameters
+
+        num_correlation_columns = num_rows_columns * (num_data_columns - 1) * num_data_columns // 2
+
+        return num_correlation_columns + num_rows_columns + num_parameters_columns
+
+    @classmethod
+    def _estimate_columns_traversal(cls, metadata, table_name,
+                                    columns_per_table, visited, distributions=None):
+        """Given a table, estimate how many columns each parent will model.
+
+        This method recursively models the children of a table all the way to the leaf nodes.
+
+        Args:
+            table_name (str):
+                Name of the table to estimate the number of columns for.
+            columns_per_table (dict):
+                Dict that stores the number of data columns + extended columns for each table.
+            visited (set):
+                Set of table names that have already been visited.
+        """
+        for child_name in metadata._get_child_map()[table_name]:
+            if child_name not in visited:
+                cls._estimate_columns_traversal(metadata, child_name, columns_per_table, visited)
+
+            columns_per_table[table_name] += \
+                cls._get_num_extended_columns(
+                    metadata, child_name, table_name, columns_per_table, distributions
+                )
+
+        visited.add(table_name)
+
+    @classmethod
+    def _estimate_num_columns(cls, metadata, distributions=None):
+        """Estimate the number of columns that will be modeled for each table.
+
+        This method estimates how many extended columns will be generated during the
+        `_augment_tables` method, so it traverses the graph in the same way.
+        If that method is ever changed, this should be updated to match.
+
+        After running this method, `columns_per_table` will store an estimate of the
+        total number of columns that each table has after running `_augment_tables`,
+        that is, the number of extended columns generated by the child tables as well
+        as the number of data columns in the table itself. `id` columns, like foreign
+        and primary keys, are not counted since they are not modeled.
+
+        Returns:
+            dict:
+                Dictionary of (table_name: int) mappings, indicating the estimated
+                number of columns that will be modeled for each table.
+        """
+        # This dict will store the number of data columns + extended columns for each table
+        # Initialize it with the number of data columns per table
+        columns_per_table = cls._get_num_data_columns(metadata)
+
+        # Starting at root tables, recursively estimate the number of columns
+        # each table will model
+        visited = set()
+        for table_name in cls._get_root_tables(metadata.relationships):
+            cls._estimate_columns_traversal(
+                metadata, table_name, columns_per_table, visited, distributions
+            )
+
+        return columns_per_table
+
     def __init__(self, metadata, locales=['en_US'], verbose=True):
         BaseMultiTableSynthesizer.__init__(self, metadata, locales=locales)
         self._table_sizes = {}
@@ -94,11 +214,21 @@ class HMASynthesizer(BaseHierarchicalSampler, BaseMultiTableSynthesizer):
 
         return super().get_learned_distributions(table_name)
 
+    def _get_distributions(self):
+        distributions = {}
+        for table in self.metadata.tables:
+            parameters = self.get_table_parameters(table)
+            sythesizer_parameter = parameters.get('synthesizer_parameters', {})
+            distributions[table] = sythesizer_parameter.get('default_distribution', None)
+
+        return distributions
+
     def _print_estimate_warning(self):
         total_est_cols = 0
-        metadata_columns = self._get_num_data_columns()
+        metadata_columns = self._get_num_data_columns(self.metadata)
         print_table = []
-        for table, est_cols in self._estimate_num_columns().items():
+        distributions = self._get_distributions()
+        for table, est_cols in self._estimate_num_columns(self.metadata, distributions).items():
             entry = []
             entry.append(table)
             entry.append(metadata_columns[table])
