+++ sdv/multi_table/utils.py
@@ -4,6 +4,8 @@ from copy import deepcopy
 
 import numpy as np
 
+from sdv.multi_table import HMASynthesizer
+
 MODELABLE_SDTYPE = ['categorical', 'numerical', 'datetime', 'boolean']
 
 
@@ -19,97 +21,6 @@ def _get_relationship_for_child(relationships, child_table):
 
 def _get_relationship_for_parent(relationships, parent_table):
     return [rel for rel in relationships if rel['parent_table_name'] == parent_table]
-
-
-def _get_num_data_columns(metadata):
-    """Get the number of data columns, ie colums that are not id, for each table.
-
-    Args:
-        metadata (MultiTableMetadata):
-            Metadata of the datasets.
-    """
-    columns_per_table = {}
-    for table_name, table in metadata.tables.items():
-        columns_per_table[table_name] = \
-            sum([1 for col in table.columns.values() if col['sdtype'] != 'id'])
-
-    return columns_per_table
-
-
-def _get_num_extended_columns(metadata, table_name, parent_table, columns_per_table):
-    """Get the number of columns that will be generated for table_name.
-
-    A table generates, for each foreign key:
-        - 1 num_rows column
-        - n*(n-1)/2 correlation columns for each data column
-        - 4 parameter columns for each data column because we are using a beta distribution
-    """
-    num_rows_columns = len(metadata._get_foreign_keys(parent_table, table_name))
-
-    # no parameter columns are generated if there are no data columns
-    num_data_columns = columns_per_table[table_name]
-    if num_data_columns == 0:
-        return num_rows_columns
-
-    num_parameters_columns = num_rows_columns * num_data_columns * 4
-
-    num_correlation_columns = num_rows_columns * (num_data_columns - 1) * num_data_columns // 2
-
-    return num_correlation_columns + num_rows_columns + num_parameters_columns
-
-
-def _estimate_columns_traversal(metadata, table_name, columns_per_table, visited):
-    """Given a table, estimate how many columns each parent will model.
-
-    This method recursively models the children of a table all the way to the leaf nodes.
-
-    Args:
-        table_name (str):
-            Name of the table to estimate the number of columns for.
-        columns_per_table (dict):
-            Dict that stores the number of data columns + extended columns for each table.
-        visited (set):
-            Set of table names that have already been visited.
-    """
-    for child_name in metadata._get_child_map()[table_name]:
-        if child_name not in visited:
-            _estimate_columns_traversal(metadata, child_name, columns_per_table, visited)
-
-        columns_per_table[table_name] += \
-            _get_num_extended_columns(metadata, child_name, table_name, columns_per_table)
-
-    visited.add(table_name)
-
-
-def _estimate_num_columns(metadata):
-    """Estimate the number of columns that will be modeled for each table.
-
-    This method estimates how many extended columns will be generated during the
-    `_augment_tables` method, so it traverses the graph in the same way.
-    If that method is ever changed, this should be updated to match.
-
-    After running this method, `columns_per_table` will store an estimate of the
-    total number of columns that each table has after running `_augment_tables`,
-    that is, the number of extended columns generated by the child tables as well
-    as the number of data columns in the table itself. `id` columns, like foreign
-    and primary keys, are not counted since they are not modeled.
-
-    Returns:
-        dict:
-            Dictionary of (table_name: int) mappings, indicating the estimated
-            number of columns that will be modeled for each table.
-    """
-    # This dict will store the number of data columns + extended columns for each table
-    # Initialize it with the number of data columns per table
-    columns_per_table = _get_num_data_columns(metadata)
-
-    # Starting at root tables, recursively estimate the number of columns
-    # each table will model
-    visited = set()
-    for table_name in _get_root_tables(metadata.relationships):
-        _estimate_columns_traversal(metadata, table_name, columns_per_table, visited)
-
-    return columns_per_table
 
 
 def _get_n_order_descendants(relationships, parent_table, order):
@@ -152,8 +63,8 @@ def _get_all_descendant_per_root_at_order_n(relationships, order):
     for root in root_tables:
         order_descendances = _get_n_order_descendants(relationships, root, order)
         all_descendant_root = set()
-        for order in order_descendances:
-            all_descendant_root.update(order_descendances[order])
+        for order_key in order_descendances:
+            all_descendant_root.update(order_descendances[order_key])
 
         all_descendants[root] = all_descendant_root
 
@@ -335,7 +246,7 @@ def _simplify_childs(metadata, childs, root_table, num_data_column):
     relationships = metadata.relationships
     max_col_per_relationships = 1000 // len(relationships)
     for child_table in childs:
-        estimate_column = _get_num_extended_columns(
+        estimate_column = HMASynthesizer._get_num_extended_columns(
             metadata, child_table, root_table, num_data_column
         )
         num_relationship = len(_get_relationship_for_child(relationships, child_table))
@@ -349,7 +260,7 @@ def _simplify_childs(metadata, childs, root_table, num_data_column):
 
 def _get_total_estimated_columns(metadata):
     """Get the total number of estimated columns from the metadata."""
-    estimate_columns_per_table = _estimate_num_columns(metadata)
+    estimate_columns_per_table = HMASynthesizer._estimate_num_columns(metadata)
     return sum(estimate_columns_per_table.values())
 
 
@@ -372,7 +283,7 @@ def _simplify_metadata(metadata):
     """
     simplify_metadata = deepcopy(metadata)
     relationships = simplify_metadata.relationships
-    all_descendants_per_root = _get_all_descendant_per_root_at_order_n(relationships, 2)
+    all_descendants_per_root = _get_all_descendant_per_root_at_order_n(relationships, order=2)
     root_with_max_descendants = max(
         all_descendants_per_root, key=lambda x: len(all_descendants_per_root[x])
     )
@@ -389,12 +300,12 @@ def _simplify_metadata(metadata):
     if grandchilds:
         simplify_metadata = _simplify_grandchilds(simplify_metadata, grandchilds)
 
-    estimate_columns_per_table = _estimate_num_columns(simplify_metadata)
+    estimate_columns_per_table = HMASynthesizer._estimate_num_columns(simplify_metadata)
     total_est_column = sum(estimate_columns_per_table.values())
     if total_est_column < 1000:
         return simplify_metadata
 
-    num_data_column = _get_num_data_columns(simplify_metadata)
+    num_data_column = HMASynthesizer._get_num_data_columns(simplify_metadata)
     simplify_metadata = _simplify_childs(
         simplify_metadata, childs, root_with_max_descendants, num_data_column
     )
