+++ tests/unit/multi_table/test_hma.py
@@ -85,10 +85,12 @@ class TestHMASynthesizer:
 
         pd.testing.assert_frame_equal(result, expected)
 
-    def test__print_estimate_warning(self, capsys):
+    @patch('sdv.multi_table.hma.HMASynthesizer._estimate_num_columns')
+    def test__print_estimate_warning(self, estimate_mock, capsys):
         """Test that a warning appears if there are more than 1000 expected columns"""
         # Setup
         metadata = get_multi_table_metadata()
+        estimate_mock.side_effect = [{'nesreca': 2000}, {'nesreca': 10}]
 
         key_phrases = [
             r'PerformanceAlert:',
@@ -97,11 +99,7 @@ class TestHMASynthesizer:
         ]
 
         # Run
-        with patch.object(HMASynthesizer,
-                          '_estimate_num_columns',
-                          return_value={'nesreca': 2000}):
-            HMASynthesizer(metadata)
-
+        HMASynthesizer(metadata)
         captured = capsys.readouterr()
 
         # Assert
@@ -110,11 +108,7 @@ class TestHMASynthesizer:
             assert match is not None
 
         # Run
-        with patch.object(HMASynthesizer,
-                          '_estimate_num_columns',
-                          return_value={'nesreca': 10}):
-            HMASynthesizer(metadata)
-
+        HMASynthesizer(metadata)
         captured = capsys.readouterr()
 
         # Assert that small amount of columns don't trigger the message
@@ -805,9 +799,10 @@ class TestHMASynthesizer:
             table_parameters={'default_distribution': 'uniform'}
         )
         synthesizer._finalize = Mock()
+        distributions = synthesizer._get_distributions()
 
         # Run estimation
-        estimated_num_columns = synthesizer._estimate_num_columns()
+        estimated_num_columns = synthesizer._estimate_num_columns(metadata, distributions)
 
         # Run actual modeling
         synthesizer.fit(data)
@@ -936,7 +931,7 @@ class TestHMASynthesizer:
         synthesizer._finalize = Mock()
 
         # Run estimation
-        estimated_num_columns = synthesizer._estimate_num_columns()
+        estimated_num_columns = synthesizer._estimate_num_columns(metadata)
 
         # Run actual modeling
         synthesizer.fit(data)
@@ -1051,7 +1046,7 @@ class TestHMASynthesizer:
         synthesizer._finalize = Mock()
 
         # Run estimation
-        estimated_num_columns = synthesizer._estimate_num_columns()
+        estimated_num_columns = synthesizer._estimate_num_columns(metadata)
 
         # Run actual modeling
         synthesizer.fit(data)
